+++ ScheduleProjectV2/main.cpp
@@ -5,7 +5,7 @@
 using namespace std;
 
 bool hasNext(string& input) {
-	for (int i = 1; i < input.length(); i++) {
+	for (size_t i = 1; i < input.length(); i++) {
 		if (input.at(i) != 32) {
 			if (input.at(i) > 0) {
 				return true;
